+++ android/gradle.properties
@@ -6,3 +6,4 @@ RNVideo_ndkversion=21.4.7075529
 RNVideo_buildToolsVersion=30.0.2
 RNVideo_media3Version=1.1.1
 RNVideo_RNVUseExoplayerIMA=false
+RNVideo_androidxCoreVersion=1.9.0
