+++ examples/basic/android/app/build.gradle
@@ -101,7 +101,11 @@ android {
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
         }
     }
+    configurations.all {
+        resolutionStrategy { force 'androidx.core:core:1.9.0' }
+    }
 }
+
 
 dependencies {
     // The version of react-native is set by the React Native Gradle Plugin
@@ -120,6 +124,15 @@ dependencies {
     }
 
     implementation project(':react-native-video')
+
+    constraints {
+        implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.8.0") {
+            because("kotlin-stdlib-jdk7 is now a part of kotlin-stdlib")
+        }
+        implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.0") {
+            because("kotlin-stdlib-jdk8 is now a part of kotlin-stdlib")
+        }
+    }
 }
 
 apply from: file("../../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesAppBuildGradle(project)
