+++ examples/basic/android/build.gradle
@@ -6,10 +6,12 @@ buildscript {
         minSdkVersion = 21
         compileSdkVersion = 33
         targetSdkVersion = 33
-        kotlinVersion = "1.6.20"
+        kotlinVersion = "1.8.0"
 
         // We use NDK 23 which has both M1 support and is the side-by-side NDK version from AGP.
         ndkVersion = "23.1.7779620"
+
+        RNVUseExoplayerIMA = true
     }
     repositories {
         google()
