+++ examples/basic/package.json
@@ -14,7 +14,6 @@
     "@react-native-picker/picker": "^1.9.11",
     "react": "18.2.0",
     "react-native": "0.72.5",
-    "react-native-video": "../../",
     "react-native-windows": "0.63.41"
   },
   "devDependencies": {
